+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -93,5 +93,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-09-05T07:11:34.253Z")
+    created: Optional[datetime] = parse_datetime("2025-09-05T09:48:59.769Z")
     r"""The date and time the resource was created"""