+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetdatapointsop.py
@@ -684,7 +684,7 @@ class ListDatasetDatapointsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-05T07:11:30.678Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-05T09:48:56.506Z")
     r"""The date and time the resource was last updated"""