+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K4CBEWR5FFP7A4EFB9KWQ9ZP"
+        "01K4CMF4H5HPD36SJGZEB5HNF7"
     )
     r"""The id of the resource"""