+++ packages/orq-rc/RELEASES.md
@@ -2649,3 +2649,13 @@ Based on:
 - [python v3.12.0-rc.41] packages/orq-rc
 ### Releases
 - [PyPI v3.12.0-rc.41] https://pypi.org/project/orq-ai-sdk/3.12.0-rc.41 - packages/orq-rc
+
+## 2025-09-05 09:50:39
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.609.0 (2.692.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.12.0-rc.42] packages/orq-rc
+### Releases
+- [PyPI v3.12.0-rc.42] https://pypi.org/project/orq-ai-sdk/3.12.0-rc.42 - packages/orq-rc